pyproject.toml
[tool.poetry]
name = "atomic-red-team-client"
version = "0.1.0"
description = "An API client for the Atomic Red Team framework"
authors = ["Fisher, Tyler <Tyler.Fisher@canadalife.com>"]
readme = "README.md"
packages = [
    { include = "atomic_red_team" },
]

[tool.poetry.scripts]
atomic-red-team = "atomic_red_team.cli:main"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
